backend/requirements.txt
bcrypt==4.1.3
blinker==1.9.0
click==8.3.0
dnspython==2.8.0
email-validator==2.1.0.post1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
google-generativeai==0.8.3
greenlet==3.2.4
gunicorn==21.2.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
Pillow==10.4.0
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.36
typing_extensions==4.15.0
Werkzeug==2.3.7
